% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-webr-repl-link.R
\name{build_webr_url}
\alias{build_webr_url}
\title{Build WebR URL with mode and code}
\usage{
build_webr_url(base_url, encoded_data, flags, mode = NULL)
}
\arguments{
\item{base_url}{Base WebR URL}

\item{encoded_data}{Base64 encoded data}

\item{flags}{URL flags}

\item{mode}{Mode string or NULL}
}
\value{
Complete WebR URL
}
\description{
Build WebR URL with mode and code
}
\keyword{internal}
