% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_character_vector}
\alias{check_character_vector}
\title{Ensure input is a character vector}
\usage{
check_character_vector(x, arg_name)
}
\arguments{
\item{x}{Object to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure input is a character vector
}
\keyword{internal}
