% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_path}
\alias{check_valid_path}
\title{Ensure input is a valid file path}
\usage{
check_valid_path(x, arg_name)
}
\arguments{
\item{x}{Object to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure input is a valid file path
}
\keyword{internal}
