% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_shinylive_engine}
\alias{check_valid_shinylive_engine}
\title{Ensure engine is valid for Shinylive}
\usage{
check_valid_shinylive_engine(engine, arg_name = "engine")
}
\arguments{
\item{engine}{Engine specification to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure engine is valid for Shinylive
}
\keyword{internal}
