% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_shinylive_mode}
\alias{check_valid_shinylive_mode}
\title{Ensure Shinylive mode is valid}
\usage{
check_valid_shinylive_mode(mode, arg_name = "mode")
}
\arguments{
\item{mode}{Mode specification to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure Shinylive mode is valid
}
\keyword{internal}
