% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_shinylive_url}
\alias{check_valid_shinylive_url}
\title{Ensure URL is a valid Shinylive URL}
\usage{
check_valid_shinylive_url(url, arg_name = "url")
}
\arguments{
\item{url}{URL to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure URL is a valid Shinylive URL
}
\keyword{internal}
