% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_version}
\alias{check_valid_version}
\title{Check if webR version is valid}
\usage{
check_valid_version(version, arg_name = "version")
}
\arguments{
\item{version}{Version string to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Check if webR version is valid
}
\details{
This function checks if the provided version is either "latest" or a valid
version string in the format "vX.Y.Z" where X, Y, Z are integers and the version is
greater than or equal to v0.5.4.
}
\keyword{internal}
