% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_webr_flags}
\alias{check_valid_webr_flags}
\title{Ensure webR flags are valid}
\usage{
check_valid_webr_flags(flags, arg_name = "flags")
}
\arguments{
\item{flags}{Flags string to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure webR flags are valid
}
\keyword{internal}
