% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{check_valid_webr_version_string}
\alias{check_valid_webr_version_string}
\title{Ensure webR version string is valid}
\usage{
check_valid_webr_version_string(version, arg_name = "version")
}
\arguments{
\item{version}{Version string to check}

\item{arg_name}{Name of the argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure webR version string is valid
}
\keyword{internal}
