% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{convert_keyvalue_to_list}
\alias{convert_keyvalue_to_list}
\title{Convert key-value structure to named list}
\usage{
convert_keyvalue_to_list(kv_item)
}
\arguments{
\item{kv_item}{Item with key and value lists}
}
\value{
Named list
}
\description{
Convert key-value structure to named list
}
\keyword{internal}
