% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{decode_multiple_webr_links}
\alias{decode_multiple_webr_links}
\title{Decode multiple webR REPL links}
\usage{
decode_multiple_webr_links(
  urls,
  output_dir,
  overwrite,
  create_subdir,
  name_dirs
)
}
\arguments{
\item{urls}{Vector of URL strings}

\item{output_dir}{Base output directory}

\item{overwrite}{Whether to overwrite files}

\item{create_subdir}{Whether to create subdirectories}

\item{name_dirs}{Whether to use numbered directory names}
}
\value{
webr_decoded_batch object
}
\description{
Decode multiple webR REPL links
}
\keyword{internal}
