% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-shinylive-link.R
\name{decode_single_shinylive_link}
\alias{decode_single_shinylive_link}
\title{Decode a single Shinylive link}
\usage{
decode_single_shinylive_link(url, output_dir, overwrite, create_subdir)
}
\arguments{
\item{url}{Single URL string}

\item{output_dir}{Output directory}

\item{overwrite}{Whether to overwrite files}

\item{create_subdir}{Whether to create subdirectory}
}
\value{
shinylive_decoded object
}
\description{
Decode a single Shinylive link
}
\keyword{internal}
