% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{decode_single_webr_link}
\alias{decode_single_webr_link}
\title{Decode a single webR REPL link}
\usage{
decode_single_webr_link(url, output_dir, overwrite, create_subdir)
}
\arguments{
\item{url}{Single URL string}

\item{output_dir}{Output directory}

\item{overwrite}{Whether to overwrite files}

\item{create_subdir}{Whether to create subdirectory}
}
\value{
webr_decoded object
}
\description{
Decode a single webR REPL link
}
\keyword{internal}
