% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-shinylive-link.R
\name{decompress_shinylive_url}
\alias{decompress_shinylive_url}
\title{Decompress and parse Shinylive URL data}
\usage{
decompress_shinylive_url(url)
}
\arguments{
\item{url}{Shinylive URL}
}
\value{
List containing engine, mode, and files_data
}
\description{
Decompress and parse Shinylive URL data
}
\keyword{internal}
