% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{decompress_webr_url}
\alias{decompress_webr_url}
\title{Decompress and parse webR URL data}
\usage{
decompress_webr_url(url)
}
\arguments{
\item{url}{webR URL}
}
\value{
List containing mode, version, flags, and files_data
}
\description{
Decompress and parse webR URL data
}
\keyword{internal}
