% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{detect_binary_content}
\alias{detect_binary_content}
\title{Detect if content is likely binary}
\usage{
detect_binary_content(content)
}
\arguments{
\item{content}{Character string to check}
}
\value{
Logical indicating if content is likely binary
}
\description{
Detect if content is likely binary
}
\keyword{internal}
