% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{ensure_files_in_list}
\alias{ensure_files_in_list}
\title{Ensure specified files exist in a named list}
\usage{
ensure_files_in_list(
  files_to_check,
  files_list,
  files_arg_name = "files",
  check_arg_name = "autorun_files"
)
}
\arguments{
\item{files_to_check}{Character vector of filenames to check}

\item{files_list}{Named list of files}

\item{files_arg_name}{Name of the files argument (for error messages)}

\item{check_arg_name}{Name of the checking argument (for error messages)}
}
\value{
Invisible TRUE if valid, aborts if not
}
\description{
Ensure specified files exist in a named list
}
\keyword{internal}
