% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-shinylive-link.R
\name{extract_code_parameter}
\alias{extract_code_parameter}
\title{Extract code parameter from Shinylive URL fragment}
\usage{
extract_code_parameter(params_string)
}
\arguments{
\item{params_string}{URL fragment part after '#'}
}
\value{
Encoded data string or NULL if not found
}
\description{
Extract code parameter from Shinylive URL fragment
}
\keyword{internal}
