% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{extract_mode_from_url}
\alias{extract_mode_from_url}
\title{Extract mode from webR URL}
\usage{
extract_mode_from_url(base_url)
}
\arguments{
\item{base_url}{Base part of URL before fragment}
}
\value{
Mode string or NULL
}
\description{
Extract mode from webR URL
}
\keyword{internal}
