% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{extract_version_from_url}
\alias{extract_version_from_url}
\title{Extract version from webR URL}
\usage{
extract_version_from_url(url)
}
\arguments{
\item{url}{webR URL}
}
\value{
Version string
}
\description{
Extract version from webR URL
}
\keyword{internal}
