% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{extract_webr_parameters}
\alias{extract_webr_parameters}
\title{Extract parameters from webR URL fragment}
\usage{
extract_webr_parameters(fragment)
}
\arguments{
\item{fragment}{URL fragment part after '#'}
}
\value{
List with code and flags
}
\description{
Extract parameters from webR URL fragment
}
\keyword{internal}
