% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinylive-link.R
\name{files_to_shinylive_json}
\alias{files_to_shinylive_json}
\title{Convert files to Shinylive JSON format}
\usage{
files_to_shinylive_json(files)
}
\arguments{
\item{files}{Named list where names are filenames and values are content}
}
\value{
List in Shinylive JSON format
}
\description{
Convert files to Shinylive JSON format
}
\keyword{internal}
