% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_file_size}
\alias{format_file_size}
\title{Format file size in human readable format}
\usage{
format_file_size(size_bytes)
}
\arguments{
\item{size_bytes}{Size in bytes}
}
\value{
Formatted string
}
\description{
Format file size in human readable format
}
\keyword{internal}
