% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{format_mode_for_display}
\alias{format_mode_for_display}
\title{Format mode string for display using cli}
\usage{
format_mode_for_display(mode)
}
\arguments{
\item{mode}{Mode string (like "editor-plot-terminal")}
}
\value{
Formatted string for display
}
\description{
Format mode string for display using cli
}
\keyword{internal}
