% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{has_directory}
\alias{has_directory}
\title{Check if directory exists}
\usage{
has_directory(path)
}
\arguments{
\item{path}{Directory path to check}
}
\value{
Logical value
}
\description{
Check if directory exists
}
\keyword{internal}
