% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{has_files_in_list}
\alias{has_files_in_list}
\title{Check if specified files exist in a named list}
\usage{
has_files_in_list(files_to_check, files_list)
}
\arguments{
\item{files_to_check}{Character vector of filenames to check}

\item{files_list}{Named list of files}
}
\value{
Logical value
}
\description{
Check if specified files exist in a named list
}
\keyword{internal}
