% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_character_vector}
\alias{is_character_vector}
\title{Check if input is a character vector}
\usage{
is_character_vector(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value
}
\description{
Check if input is a character vector
}
\keyword{internal}
