% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_named_list}
\alias{is_named_list}
\title{Check if input is a properly named list}
\usage{
is_named_list(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value
}
\description{
Check if input is a properly named list
}
\keyword{internal}
