% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_single_logical}
\alias{is_single_logical}
\title{Check if input is a single logical value}
\usage{
is_single_logical(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value
}
\description{
Check if input is a single logical value
}
\keyword{internal}
