% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_single_string}
\alias{is_single_string}
\title{Check if input is a single character string}
\usage{
is_single_string(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value
}
\description{
Check if input is a single character string
}
\keyword{internal}
