% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_mode}
\alias{is_valid_mode}
\title{Check if mode is valid}
\usage{
is_valid_mode(mode)
}
\arguments{
\item{mode}{Mode specification (character vector or string)}
}
\value{
Logical value
}
\description{
Check if mode is valid
}
\keyword{internal}
