% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_path}
\alias{is_valid_path}
\title{Check if input is a valid file path}
\usage{
is_valid_path(x)
}
\arguments{
\item{x}{Object to check}
}
\value{
Logical value
}
\description{
Check if input is a valid file path
}
\keyword{internal}
