% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_shinylive_engine}
\alias{is_valid_shinylive_engine}
\title{Create Shinylive sharelinks for R and Python Shiny apps}
\usage{
is_valid_shinylive_engine(engine)
}
\arguments{
\item{engine}{Engine specification ("r" or "python")}
}
\value{
Logical value
}
\description{
Functions to create Shinylive links for R and Python Shiny applications.
Shinylive allows running Shiny apps entirely in the browser.
Check if engine is valid for Shinylive
}
\keyword{internal}
