% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_shinylive_mode}
\alias{is_valid_shinylive_mode}
\title{Check if Shinylive mode is valid}
\usage{
is_valid_shinylive_mode(mode)
}
\arguments{
\item{mode}{Mode specification ("editor" or "app")}
}
\value{
Logical value
}
\description{
Check if Shinylive mode is valid
}
\keyword{internal}
