% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_shinylive_url}
\alias{is_valid_shinylive_url}
\title{Check if URL is a valid Shinylive URL}
\usage{
is_valid_shinylive_url(url)
}
\arguments{
\item{url}{URL string to check}
}
\value{
Logical value
}
\description{
Check if URL is a valid Shinylive URL
}
\keyword{internal}
