% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_webr_flags}
\alias{is_valid_webr_flags}
\title{Check if webR URL has valid flags}
\usage{
is_valid_webr_flags(flags)
}
\arguments{
\item{flags}{Flags string to check}
}
\value{
Logical value
}
\description{
Check if webR URL has valid flags
}
\keyword{internal}
