% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_webr_url}
\alias{is_valid_webr_url}
\title{Check if URL is a valid webR URL}
\usage{
is_valid_webr_url(url)
}
\arguments{
\item{url}{URL string to check}
}
\value{
Logical value
}
\description{
Check if URL is a valid webR URL
}
\keyword{internal}
