% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validators.R
\name{is_valid_webr_version_string}
\alias{is_valid_webr_version_string}
\title{Check if webR version string is valid}
\usage{
is_valid_webr_version_string(version)
}
\arguments{
\item{version}{Version string to check}
}
\value{
Logical value
}
\description{
Check if webR version string is valid
}
\keyword{internal}
