% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shinylive_decoded}
\alias{new_shinylive_decoded}
\title{Create a shinylive_decoded object}
\usage{
new_shinylive_decoded(files_info, output_dir, url, engine, mode)
}
\arguments{
\item{files_info}{Data frame with file information}

\item{output_dir}{Directory where files were saved}

\item{url}{Original Shinylive URL}

\item{engine}{Engine type}

\item{mode}{Shinylive mode}
}
\value{
shinylive_decoded object
}
\description{
Create a shinylive_decoded object
}
\keyword{internal}
