% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shinylive_decoded_batch}
\alias{new_shinylive_decoded_batch}
\title{Create a shinylive_decoded_batch object}
\usage{
new_shinylive_decoded_batch(results, base_dir, urls)
}
\arguments{
\item{results}{List of shinylive_decoded objects}

\item{base_dir}{Base output directory}

\item{urls}{Original URLs}
}
\value{
shinylive_decoded_batch object
}
\description{
Create a shinylive_decoded_batch object
}
\keyword{internal}
