% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shinylive_directory}
\alias{new_shinylive_directory}
\title{Create a shinylive_directory object}
\usage{
new_shinylive_directory(urls, engine, mode, source_directory)
}
\arguments{
\item{urls}{Named character vector of URLs}

\item{engine}{Engine type ("r" or "python")}

\item{mode}{Shinylive mode ("editor" or "app")}

\item{source_directory}{Original source directory}
}
\value{
shinylive_directory object
}
\description{
Create a shinylive_directory object
}
\keyword{internal}
