% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shinylive_preview}
\alias{new_shinylive_preview}
\title{Create a shinylive_preview object}
\usage{
new_shinylive_preview(url, engine, mode, files_data, total_size, file_types)
}
\arguments{
\item{url}{Original Shinylive URL}

\item{engine}{Engine type}

\item{mode}{Shinylive mode}

\item{files_data}{Raw file data}

\item{total_size}{Total size in bytes}

\item{file_types}{Unique file types found}
}
\value{
shinylive_preview object
}
\description{
Create a shinylive_preview object
}
\keyword{internal}
