% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_shinylive_project}
\alias{new_shinylive_project}
\title{Create a shinylive_project object}
\usage{
new_shinylive_project(url, files, engine, mode)
}
\arguments{
\item{url}{Shinylive URL}

\item{files}{Named list of file contents}

\item{engine}{Engine type ("r" or "python")}

\item{mode}{Shinylive mode ("editor" or "app")}
}
\value{
shinylive_project object
}
\description{
Create a shinylive_project object
}
\keyword{internal}
