% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_decoded}
\alias{new_webr_decoded}
\title{Create a webr_decoded object}
\usage{
new_webr_decoded(files_info, output_dir, url, mode, version, flags)
}
\arguments{
\item{files_info}{Data frame with file information}

\item{output_dir}{Output directory path}

\item{url}{Original webR URL}

\item{mode}{Interface mode}

\item{version}{WebR version}

\item{flags}{Encoding flags}
}
\value{
webr_decoded object
}
\description{
Create a webr_decoded object
}
\keyword{internal}
