% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_decoded_batch}
\alias{new_webr_decoded_batch}
\title{Create a webr_decoded_batch object}
\usage{
new_webr_decoded_batch(results, base_output_dir, urls)
}
\arguments{
\item{results}{List of webr_decoded objects}

\item{base_output_dir}{Base output directory}

\item{urls}{Original URLs}
}
\value{
webr_decoded_batch object
}
\description{
Create a webr_decoded_batch object
}
\keyword{internal}
