% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_directory}
\alias{new_webr_directory}
\title{Create a webr_directory object}
\usage{
new_webr_directory(urls, base_path, mode, version, source_directory)
}
\arguments{
\item{urls}{Named character vector of URLs}

\item{base_path}{Base directory path in WebR}

\item{mode}{Interface mode}

\item{version}{WebR version}

\item{source_directory}{Original source directory}
}
\value{
webr_directory object
}
\description{
Create a webr_directory object
}
\keyword{internal}
