% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_exercise}
\alias{new_webr_exercise}
\title{Create a webr_exercise object}
\usage{
new_webr_exercise(exercise, solution)
}
\arguments{
\item{exercise}{webr_link object for exercise}

\item{solution}{webr_link object for solution}
}
\value{
webr_exercise object
}
\description{
Create a webr_exercise object
}
\keyword{internal}
