% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_link}
\alias{new_webr_link}
\title{Create a webr_link object}
\usage{
new_webr_link(url, filename, path, mode, version, autorun)
}
\arguments{
\item{url}{WebR URL}

\item{filename}{Original filename}

\item{path}{Full path in WebR}

\item{mode}{Interface mode}

\item{version}{WebR version}

\item{autorun}{Whether autorun is enabled}
}
\value{
webr_link object
}
\description{
Create a webr_link object
}
\keyword{internal}
