% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_preview}
\alias{new_webr_preview}
\title{Create a webr_preview object}
\usage{
new_webr_preview(
  url,
  mode,
  version,
  flags,
  files_data,
  total_size,
  autorun_files
)
}
\arguments{
\item{url}{Original webR URL}

\item{mode}{Interface mode}

\item{version}{WebR version}

\item{flags}{Encoding flags}

\item{files_data}{Raw file data}

\item{total_size}{Total size in bytes}

\item{autorun_files}{Files with autorun enabled}
}
\value{
webr_preview object
}
\description{
Create a webr_preview object
}
\keyword{internal}
