% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_webr_project}
\alias{new_webr_project}
\title{Create a webr_project object}
\usage{
new_webr_project(url, files, base_path, mode, version, autorun_files)
}
\arguments{
\item{url}{WebR URL}

\item{files}{Named list of file contents}

\item{base_path}{Base directory path}

\item{mode}{Interface mode}

\item{version}{WebR version}

\item{autorun_files}{Files with autorun enabled}
}
\value{
webr_project object
}
\description{
Create a webr_project object
}
\keyword{internal}
