% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{normalize_msgpack_data}
\alias{normalize_msgpack_data}
\title{Normalize msgpack data from RcppMsgPack format to list format}
\usage{
normalize_msgpack_data(msgpack_data)
}
\arguments{
\item{msgpack_data}{Raw msgpack data from RcppMsgPack}
}
\value{
Normalized list structure
}
\description{
Normalize msgpack data from RcppMsgPack format to list format
}
\keyword{internal}
