% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-webr-repl-link.R
\name{parse_webr_url}
\alias{parse_webr_url}
\title{Parse webR URL structure}
\usage{
parse_webr_url(url)
}
\arguments{
\item{url}{webR URL to parse}
}
\value{
List with URL components
}
\description{
Parse webR URL structure
}
\keyword{internal}
