% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_decoded}
\alias{print.shinylive_decoded}
\title{Print method for shinylive_decoded objects}
\usage{
\method{print}{shinylive_decoded}(x, ...)
}
\arguments{
\item{x}{shinylive_decoded object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_decoded objects
}
