% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_preview}
\alias{print.shinylive_preview}
\title{Print method for shinylive_preview objects}
\usage{
\method{print}{shinylive_preview}(x, show_content = FALSE, max_content_length = 500, ...)
}
\arguments{
\item{x}{shinylive_preview object}

\item{show_content}{Logical. Whether to print the contents of each file}

\item{max_content_length}{Maximum number of characters of content to show per file}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_preview objects
}
