% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{simple_hash}
\alias{simple_hash}
\title{Create a simple hash from a string using base R}
\usage{
simple_hash(x, length = 8)
}
\arguments{
\item{x}{String to hash}

\item{length}{Length of hash to return (default: 8)}
}
\value{
Character string hash
}
\description{
Create a simple hash from a string using base R
}
\keyword{internal}
