% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-link.R
\name{as.character.webr_link}
\alias{as.character.webr_link}
\alias{as.character.shinylive_link}
\alias{as.character.webr_project}
\alias{as.character.webr_exercise}
\alias{as.character.webr_directory}
\alias{as.character.webr_decoded}
\alias{as.character.webr_decoded_batch}
\alias{as.character.webr_preview}
\alias{as.character.shinylive_project}
\alias{as.character.shinylive_directory}
\alias{as.character.shinylive_decoded}
\alias{as.character.shinylive_decoded_batch}
\alias{as.character.shinylive_preview}
\title{Extract URLs as character vector}
\usage{
\method{as.character}{webr_link}(x, ...)

\method{as.character}{shinylive_link}(x, ...)

\method{as.character}{webr_project}(x, ...)

\method{as.character}{webr_exercise}(x, ...)

\method{as.character}{webr_directory}(x, ...)

\method{as.character}{webr_decoded}(x, ...)

\method{as.character}{webr_decoded_batch}(x, ...)

\method{as.character}{webr_preview}(x, ...)

\method{as.character}{shinylive_project}(x, ...)

\method{as.character}{shinylive_directory}(x, ...)

\method{as.character}{shinylive_decoded}(x, ...)

\method{as.character}{shinylive_decoded_batch}(x, ...)

\method{as.character}{shinylive_preview}(x, ...)
}
\arguments{
\item{x}{Link object}

\item{...}{Additional arguments}
}
\value{
A character vector of URLs.
\itemize{
\item Most objects give a single URL.
\item An exercise gives two, named \code{exercise} and \code{solution}.
\item A directory or a batch gives one URL for each file.
}
}
\description{
Pulls the shareable URL(s) out of a livelink object.
}
\seealso{
\code{\link[=repl_urls]{repl_urls()}} for the same result from a function you can call by name.
}
