% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_decoded_batch}
\alias{print.shinylive_decoded_batch}
\title{Print method for shinylive_decoded_batch objects}
\usage{
\method{print}{shinylive_decoded_batch}(x, ...)
}
\arguments{
\item{x}{shinylive_decoded_batch object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_decoded_batch objects
}
