% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_directory}
\alias{print.shinylive_directory}
\title{Print method for shinylive_directory objects}
\usage{
\method{print}{shinylive_directory}(x, ...)
}
\arguments{
\item{x}{shinylive_directory object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_directory objects
}
