% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_link}
\alias{print.shinylive_link}
\title{Print method for shinylive_link objects}
\usage{
\method{print}{shinylive_link}(x, ...)
}
\arguments{
\item{x}{shinylive_link object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_link objects
}
