% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.shinylive_project}
\alias{print.shinylive_project}
\title{Print method for shinylive_project objects}
\usage{
\method{print}{shinylive_project}(x, ...)
}
\arguments{
\item{x}{shinylive_project object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for shinylive_project objects
}
