% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.webr_decoded}
\alias{print.webr_decoded}
\title{Print method for webr_decoded objects}
\usage{
\method{print}{webr_decoded}(x, ...)
}
\arguments{
\item{x}{webr_decoded object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for webr_decoded objects
}
