% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.webr_directory}
\alias{print.webr_directory}
\title{Print method for webr_directory objects}
\usage{
\method{print}{webr_directory}(x, ...)
}
\arguments{
\item{x}{webr_directory object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for webr_directory objects
}
