% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.webr_exercise}
\alias{print.webr_exercise}
\title{Print method for webr_exercise objects}
\usage{
\method{print}{webr_exercise}(x, ...)
}
\arguments{
\item{x}{webr_exercise object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for webr_exercise objects
}
