% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.webr_link}
\alias{print.webr_link}
\title{Print method for webr_link objects}
\usage{
\method{print}{webr_link}(x, ...)
}
\arguments{
\item{x}{webr_link object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for webr_link objects
}
