% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{print.webr_project}
\alias{print.webr_project}
\title{Print method for webr_project objects}
\usage{
\method{print}{webr_project}(x, ...)
}
\arguments{
\item{x}{webr_project object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns the object
}
\description{
Print method for webr_project objects
}
