% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-link.R
\name{repl_urls}
\alias{repl_urls}
\alias{repl_urls.webr_link}
\alias{repl_urls.webr_project}
\alias{repl_urls.webr_exercise}
\alias{repl_urls.webr_directory}
\alias{repl_urls.webr_decoded}
\alias{repl_urls.webr_decoded_batch}
\alias{repl_urls.webr_preview}
\alias{repl_urls.shinylive_link}
\alias{repl_urls.shinylive_project}
\alias{repl_urls.shinylive_directory}
\alias{repl_urls.shinylive_decoded}
\alias{repl_urls.shinylive_decoded_batch}
\alias{repl_urls.default}
\title{Get WebR URLs from objects}
\usage{
repl_urls(x, ...)

\method{repl_urls}{webr_link}(x, ...)

\method{repl_urls}{webr_project}(x, ...)

\method{repl_urls}{webr_exercise}(x, ...)

\method{repl_urls}{webr_directory}(x, ...)

\method{repl_urls}{webr_decoded}(x, ...)

\method{repl_urls}{webr_decoded_batch}(x, ...)

\method{repl_urls}{webr_preview}(x, ...)

\method{repl_urls}{shinylive_link}(x, ...)

\method{repl_urls}{shinylive_project}(x, ...)

\method{repl_urls}{shinylive_directory}(x, ...)

\method{repl_urls}{shinylive_decoded}(x, ...)

\method{repl_urls}{shinylive_decoded_batch}(x, ...)

\method{repl_urls}{default}(x, ...)
}
\arguments{
\item{x}{WebR object (webr_link, webr_project, webr_exercise, or webr_directory)}

\item{...}{Additional arguments passed to methods}
}
\value{
Character vector of URLs
}
\description{
Generic function to extract WebR URLs from different object types.
Provides a clear way to get just the URLs for sharing or further processing.
}
\examples{
\dontrun{
# Single link
link <- webr_repl_link("plot(1:10)")
repl_urls(link)

# Exercise (returns named vector)
exercise <- webr_repl_exercise("# TODO", "plot(1:10)", "test")
repl_urls(exercise)

# Directory (returns named vector)
links <- webr_repl_directory("./examples/")
repl_urls(links)

# Decoded files (returns original URL)
decoded <- decode_webr_link("https://webr.r-wasm.org/latest/#code=...")
repl_urls(decoded)
}

}
